pixel-prompt-lambda/requirements.txt
Pillow
requests
huggingface_hub
transformers
sentencepiece
protobuf
aiohttp
boto3
groq
openai
python-dotenv